conflict : move 2 classes
RmesExceptionHandler and RmesRuntimeExceptionHandler
